apps/text-extractor/requirements.txt
aio-pika
redis
httpx
pdfplumber
beautifulsoup4
lxml
pydantic-settings